requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
tenacity==8.2.3
